requirements.txt
# Literature Monitor Dependencies

# Core
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database (Phase 2)
# sqlite is built-in

# Claude API (Phase 3)
anthropic>=0.18.0

# Email (Phase 4)
# smtplib is built-in

# Web UI
flask>=3.0.0

# Development
pytest>=7.0.0
